update login page, UX for the app

+++ Flutter/flutter_practice_one/apps/fitness_app/lib/data/models/user_data.dart
@@ -5,8 +5,8 @@ class UserModel {
 
   String? image;
   String? name;
-  double? weight;
-  double? height;
+  int? weight;
+  int? height;
   int? age;
 }
